chore: move vite module types to core to avoid circle reference
