+++ packages/client/package.json
@@ -61,7 +61,6 @@
     "unplugin-auto-import": "^0.17.5",
     "unplugin-vue-components": "^0.26.0",
     "vite": "^5.1.4",
-    "vite-plugin-vue-devtools": "workspace:^",
     "vue": "^3.4.21"
   }
 }
