+++ packages/core/src/vite-bridge/index.ts
@@ -1,2 +1,3 @@
 export * from './client'
 export * from './server'
+export * from './module-types'
